+++ app/controllers/district.controller.js
@@ -1,10 +1,10 @@
 const { save } = require('helpers/providers/cache/redisClient');
-const Distrit = require('models/district.model');
+const District = require('models/district.model');
 const getCaching = require('./app.controller');
 
-// Retrieve all distrits from the database.
+// Retrieve all districts from the database.
 exports.findAll = async (request, response) => {
-  const field = 'distrits';
+  const field = 'districts';
   const resultCache = await getCaching(field);
 
   if (resultCache) {
@@ -16,10 +16,10 @@ exports.findAll = async (request, response) => {
     return;
   }
 
-  Distrit.getAll((err, data) => {
+  District.getAll((err, data) => {
     if (err) {
       response.status(500).send({
-        message: err.message || 'Some error occurred while retrieving distrit.',
+        message: err.message || 'Some error occurred while retrieving district.',
       });
     } else {
       // Update cache.
@@ -36,10 +36,10 @@ exports.findAll = async (request, response) => {
 
 // Get longitude and latitude of a specific district.
 exports.getLngLat = async (request, response) => {
-  Distrit.getLngLat(request.params, (err, data) => {
+  District.getLngLat(request.params, (err, data) => {
     if (err) {
       response.status(403).send({
-        message: request.polyglot.t('not_retrieve_distrit') || err.message,
+        message: request.polyglot.t('not_retrieve_district') || err.message,
       });
     } else {
       const json = {
@@ -47,6 +47,18 @@ exports.getLngLat = async (request, response) => {
         data,
       };
       response.status(200).json(json);
+    }
+  });
+};
+
+exports.findById = (request, response) => {
+  District.findById(request.params, (err, data) => {
+    if (err) {
+      response.status(403).send({
+        message: request.polyglot.t('not_retrieve_district') || err.message,
+      });
+    } else {
+      response.status(200).json(data);
     }
   });
 };