+++ app/routes/district.routes.js
@@ -4,6 +4,7 @@ const districts = require('controllers/district.controller');
 const router = express.Router();
 
 router.get('/distritos', districts.findAll);
-router.get('/distritos/:name', districts.getLngLat);
+router.get('/distritos/:id', districts.findById);
+// router.get('/distritos/:name', districts.getLngLat);
 
 module.exports = router;